Make obfuscation key 4 x int32

+++ utils/obfuscate/ids.go
@@ -25,7 +25,7 @@ const (
 )
 
 // EncodeID encodes a numeric ID into an obfuscated string using a Feistel cipher with the provided key.
-func EncodeID(id int64, key [4]int64) (string, error) {
+func EncodeID(id int64, key [4]int32) (string, error) {
 	if id <= 0 || id > maxID {
 		return "", errors.New("encode requires id between 1 and 9,999,999,999")
 	}
@@ -51,7 +51,7 @@ func EncodeID(id int64, key [4]int64) (string, error) {
 }
 
 // DecodeID decodes an obfuscated string back into the original numeric ID using a Feistel cipher with the provided key.
-func DecodeID(code string, key [4]int64) (int64, error) {
+func DecodeID(code string, key [4]int32) (int64, error) {
 	if !WasID(code) {
 		return 0, errors.New("code is not a valid obfuscated ID")
 	}
@@ -86,49 +86,49 @@ func WasID(code string) bool {
 	return true
 }
 
-func feistel30Encrypt(n int64, key [4]int64) int64 {
+func feistel30Encrypt(n int64, key [4]int32) int64 {
 	left := (n >> f30HalfSize) & f30Mask
 	right := n & f30Mask
 
 	for _, k := range key {
 		newLeft := right
-		right = left ^ (feistelRound(right, k, f30Mask) & f30Mask)
+		right = left ^ (feistelRound(right, int64(k), f30Mask) & f30Mask)
 		left = newLeft
 	}
 	return (left << f30HalfSize) | right
 }
 
-func feistel30Decrypt(n int64, key [4]int64) int64 {
+func feistel30Decrypt(n int64, key [4]int32) int64 {
 	left := (n >> f30HalfSize) & f30Mask
 	right := n & f30Mask
 
 	for i := len(key) - 1; i >= 0; i-- {
 		newRight := left
-		left = right ^ (feistelRound(left, key[i], f30Mask) & f30Mask)
+		left = right ^ (feistelRound(left, int64(key[i]), f30Mask) & f30Mask)
 		right = newRight
 	}
 	return (left << f30HalfSize) | right
 }
 
-func feistel34Encrypt(n int64, key [4]int64) int64 {
+func feistel34Encrypt(n int64, key [4]int32) int64 {
 	left := (n >> f34HalfSize) & f34Mask
 	right := n & f34Mask
 
 	for _, k := range key {
 		newLeft := right
-		right = left ^ (feistelRound(right, k, f34Mask) & f34Mask)
+		right = left ^ (feistelRound(right, int64(k), f34Mask) & f34Mask)
 		left = newLeft
 	}
 	return (left << f34HalfSize) | right
 }
 
-func feistel34Decrypt(n int64, key [4]int64) int64 {
+func feistel34Decrypt(n int64, key [4]int32) int64 {
 	left := (n >> f34HalfSize) & f34Mask
 	right := n & f34Mask
 
 	for i := len(key) - 1; i >= 0; i-- {
 		newRight := left
-		left = right ^ (feistelRound(left, key[i], f34Mask) & f34Mask)
+		left = right ^ (feistelRound(left, int64(key[i]), f34Mask) & f34Mask)
 		right = newRight
 	}
 	return (left << f34HalfSize) | right
